+++ app/src/main/java/com/example/mangoplate_mock_aos_radi/src/main/home/model/HomeRecyclerItems.kt
@@ -2,4 +2,4 @@ package com.example.mangoplate_mock_aos_radi.src.main.home.model
 
 data class HomeRecyclerItems(var image: String, var title: String, var areaName: String,
                              var star: String, var viewPoint: Int, var reviewCount: Int,
-                             var idx: Int, var restaurantId: Int, var distanceFromUser: Int)
+                             var idx: Int, var restaurantId: Int, var distanceFromUser: Float)
