+++ app/src/main/java/com/example/mangoplate_mock_aos_radi/src/main/home/model/RestaurantResultData.kt
@@ -2,7 +2,7 @@ package com.example.mangoplate_mock_aos_radi.src.main.home.model
 
 data class RestaurantResultData (val restaurantId: Int,
                                  val restaurantName: String,
-                                 val distanceFromUser: Int,
+                                 val distanceFromUser: Float,
                                  val areaName: String,
                                  val restaurantView: Int,
                                  val reviewCount: Int,
